TCES-12 refactor(auth) refactor all controllers to be separate files
