+++ server/src/controllers/auth/createUser.js
@@ -1,4 +1,4 @@
-const User = require('../models/user.model');
+const User = require('../../models/user.model');
 const crypto = require('crypto');
 
 const createUserRequestHandler = async (req, res, next) => {
@@ -33,29 +33,4 @@ const createUserRequestHandler = async (req, res, next) => {
     }
 }
 
-const logoutRequestHandler = (req, res, next) => {
-    req.logout((err) => {
-        if (err) { return next(err); }
-        res.redirect('/');
-    });
-}
-
-const isLoggedInRequestHandler = (req, res) => {
-    if (req.user) {
-        if (req.user.is_admin) {
-            res.send(`User ${req.user.username} is logged in as admin`);
-            return;
-        }
-        res.send(`User ${req.user} logged in, but not an admin`);
-        return;
-    }
-    res.send("user not logged in");
-    return;
-}
-
-
-module.exports = {
-    createUserRequestHandler,
-    logoutRequestHandler,
-    isLoggedInRequestHandler
-}
+module.exports = createUserRequestHandler;
