+++ server/src/controllers/auth/logout.js
@@ -1,0 +1,8 @@
+const logoutRequestHandler = (req, res, next) => {
+    req.logout((err) => {
+        if (err) { return next(err); }
+        res.redirect('/');
+    });
+}
+
+module.exports = logoutRequestHandler;
